dpdp-guardian-main (1)/public/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==0.28.1
google-generativeai==0.3.2
python-dotenv==1.0.0